add patterns for I and NI

+++ src/consts.rs
@@ -126,4 +126,4 @@ pub const WIN_LENGTH: usize = 5;
 pub const MAX_DEFCON: usize = WIN_LENGTH;
 
 // Max defcon for immediate threat.
-pub const MDFIT: usize = 2;
+pub const MDFIT: i8 = 2;
